+++ Events.php
@@ -2,16 +2,19 @@
 
 namespace humhub\modules\fcmPush;
 
-
 use humhub\components\mail\Message;
+use humhub\helpers\DeviceDetectorHelper;
 use humhub\modules\fcmPush\assets\FcmPushAsset;
 use humhub\modules\fcmPush\assets\FirebaseAsset;
 use humhub\modules\fcmPush\components\MailerMessage;
 use humhub\modules\fcmPush\components\NotificationTargetProvider;
 use humhub\modules\fcmPush\helpers\MobileAppHelper;
+use humhub\modules\fcmPush\helpers\WebAppHelper;
 use humhub\modules\fcmPush\services\DriverService;
 use humhub\modules\fcmPush\widgets\PushNotificationInfoWidget;
 use humhub\modules\notification\targets\MobileTargetProvider;
+use humhub\modules\ui\menu\MenuLink;
+use humhub\modules\user\widgets\AccountTopMenu;
 use humhub\modules\user\widgets\AuthChoice;
 use humhub\modules\web\pwa\controllers\ManifestController;
 use humhub\modules\web\pwa\controllers\ServiceWorkerController;
@@ -21,10 +24,6 @@ use yii\base\Event;
 
 class Events
 {
-
-    private const SESSION_VAR_LOGOUT = 'mobileAppHandleLogout';
-    private const SESSION_VAR_LOGIN = 'mobileAppHandleLogin';
-
     public static function onBeforeRequest()
     {
         /** @var Module $module */
@@ -65,22 +64,18 @@ class Events
         // Service Worker Addons
         $controller->additionalJs .= <<<JS
             // Give the service worker access to Firebase Messaging.
-            importScripts('{$bundle->baseUrl}/firebase-app.js');
-            importScripts('{$bundle->baseUrl}/firebase-messaging.js');
-            //importScripts('https://www.gstatic.com/firebasejs/6.3.3/firebase-app.js');
-            //importScripts('https://www.gstatic.com/firebasejs/6.3.3/firebase-messaging.js');
-        
-           firebase.initializeApp({messagingSenderId: "{$pushDriver->getSenderId()}"});
-            
-            const messaging = firebase.messaging();
-            messaging.setBackgroundMessageHandler(function(payload) {
-              const notificationTitle = payload.data.title;
-              const notificationOptions = {
-                body: payload.data.body,
-                icon: payload.data.icon
-              };
-              return self.registration.showNotification(notificationTitle, notificationOptions);
+            importScripts('{$bundle->baseUrl}/firebase-app-compat.js');
+            importScripts('{$bundle->baseUrl}/firebase-messaging-compat.js');
+
+            firebase.initializeApp({
+                messagingSenderId: "{$pushDriver->getSenderId()}",
+                projectId: "{$module->getConfigureForm()->getJsonParam('project_id')}",
+                appId: "{$module->getConfigureForm()->firebaseAppId}",
+                apiKey: "{$module->getConfigureForm()->firebaseApiKey}",
             });
+
+            // Initialize Firebase Cloud Messaging and get a reference to the service
+            firebase.messaging();
 JS;
     }
 
@@ -90,27 +85,46 @@ JS;
         $baseStack = $event->sender;
 
         $baseStack->addWidget(PushNotificationInfoWidget::class);
-
     }
 
     public static function onLayoutAddonInit($event)
     {
-        if (Yii::$app->session->has(self::SESSION_VAR_LOGOUT)) {
-            MobileAppHelper::unregisterNotificationScript(); // Before Logout
-            MobileAppHelper::registerLogoutScript();
-            Yii::$app->session->remove(self::SESSION_VAR_LOGOUT);
+        // If the mobile app Opener page is open (after login and switching instance)
+        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_HIDE_OPENER)) {
+            MobileAppHelper::registerHideOpenerScript();
+            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_HIDE_OPENER);
+        } elseif (DeviceDetectorHelper::appOpenerState()) {
+            MobileAppHelper::registerHideOpenerScript();
         }
 
-        if (Yii::$app->session->has(self::SESSION_VAR_LOGIN)) {
-            MobileAppHelper::registerLoginScript();
+        // After login
+        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_REGISTER_NOTIFICATION)) {
             MobileAppHelper::registerNotificationScript();
-            Yii::$app->session->remove(self::SESSION_VAR_LOGIN);
+            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_REGISTER_NOTIFICATION);
         }
 
-        if (Yii::$app->user->isGuest) {
-            return;
+        // After logout
+        if (Yii::$app->session->has(WebAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION)) {
+            static::registerAssets();
+            WebAppHelper::unregisterNotificationScript();
+            Yii::$app->session->remove(WebAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION);
+        }
+        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION)) {
+            MobileAppHelper::unregisterNotificationScript();
+            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION);
+        }
+        if (Yii::$app->session->has(MobileAppHelper::SESSION_VAR_SHOW_OPENER)) {
+            MobileAppHelper::registerShowOpenerScript();
+            Yii::$app->session->remove(MobileAppHelper::SESSION_VAR_SHOW_OPENER);
         }
 
+        if (!Yii::$app->user->isGuest) {
+            static::registerAssets();
+        }
+    }
+
+    private static function registerAssets()
+    {
         /** @var Module $module */
         $module = Yii::$app->getModule('fcm-push');
 
@@ -120,31 +134,52 @@ JS;
 
         FcmPushAsset::register(Yii::$app->view);
         FirebaseAsset::register(Yii::$app->view);
-
-    }
-
-    public static function onAfterLogout()
-    {
-        Yii::$app->session->set(self::SESSION_VAR_LOGOUT, 1);
     }
 
     public static function onAfterLogin()
     {
-        Yii::$app->session->set(self::SESSION_VAR_LOGIN, 1);
+        Yii::$app->session->set(MobileAppHelper::SESSION_VAR_HIDE_OPENER, 1);
+        Yii::$app->session->set(MobileAppHelper::SESSION_VAR_REGISTER_NOTIFICATION, 1);
     }
 
-    public static function onAuthChoiceBeforeRun(Event $event) {
+    public static function onAfterLogout()
+    {
+        Yii::$app->session->set(WebAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
+        Yii::$app->session->set(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
+        Yii::$app->session->set(MobileAppHelper::SESSION_VAR_SHOW_OPENER, 1);
+    }
+
+    public static function onAuthChoiceBeforeRun(Event $event)
+    {
         /** @var AuthChoice $sender */
         $sender = $event->sender;
 
         /** @var Module $module */
         $module = Yii::$app->getModule('fcm-push');
 
-        if (MobileAppHelper::isIosApp() && $module->getConfigureForm()->disableAuthChoicesIos) {
+        if (DeviceDetectorHelper::isIosApp() && $module->getConfigureForm()->disableAuthChoicesIos) {
             $sender->setClients([]);
         }
-
-
     }
 
+    public static function onAccountTopMenuInit(Event $event)
+    {
+        if (!DeviceDetectorHelper::isMultiInstanceApp()) {
+            return;
+        }
+
+        /** @var AccountTopMenu $menu */
+        $menu = $event->sender;
+
+        $menu->addEntry(new MenuLink([
+            'label' => Yii::t('FcmPushModule.base', 'Switch network'),
+            'url' => ['/fcm-push/mobile-app/instance-opener'],
+            'icon' => 'arrows-h',
+            'sortOrder' => 699, // Just before "Logout"
+            'isVisible' => true,
+            'htmlOptions' => [
+                'data-pjax' => '0', // Force full page refresh to trigger the onLayoutAddonInit event
+            ],
+        ]));
+    }
 }