+++ Module.php
@@ -11,7 +11,6 @@ use yii\helpers\Url;
 class Module extends \humhub\components\Module
 {
     public string $humhubProxySenderId = '21392898126';
-
     private ?ConfigureForm $configForm = null;
     private ?DriverService $driverService = null;
     private ?GoService $goService = null;