+++ config.php
@@ -3,14 +3,15 @@
 
 /** @noinspection MissedFieldInspection */
 
-use humhub\modules\fcmPush\Events;
 use humhub\components\Controller;
-
-//use humhub\modules\notification\widgets\NotificationInfoWidget;
+use humhub\modules\fcmPush\Events;
+use humhub\modules\user\components\User;
+use humhub\modules\user\widgets\AccountTopMenu;
 use humhub\modules\user\widgets\AuthChoice;
 use humhub\widgets\LayoutAddons;
 use yii\base\Application;
-use humhub\modules\user\components\User;
+
+//use humhub\modules\notification\widgets\NotificationInfoWidget;
 
 return [
     'id' => 'fcm-push',
@@ -21,13 +22,16 @@ return [
         ['humhub\modules\web\pwa\controllers\ServiceWorkerController', Controller::EVENT_INIT, [Events::class, 'onServiceWorkerControllerInit']],
         [LayoutAddons::class, LayoutAddons::EVENT_INIT, [Events::class, 'onLayoutAddonInit']],
         [Application::class, Application::EVENT_BEFORE_REQUEST, [Events::class, 'onBeforeRequest']],
-        [User::class, User::EVENT_AFTER_LOGOUT, [Events::class, 'onAfterLogout']],
         [User::class, User::EVENT_AFTER_LOGIN, [Events::class, 'onAfterLogin']],
+        [User::class, User::EVENT_AFTER_LOGOUT, [Events::class, 'onAfterLogout']],
         [AuthChoice::class, AuthChoice::EVENT_BEFORE_RUN, [Events::class, 'onAuthChoiceBeforeRun']],
-        //[NotificationInfoWidget::class, \humhub\widgets\BaseStack::EVENT_RUN, [Events::class, 'onNotificationInfoWidget']]
+        //[NotificationInfoWidget::class, \humhub\widgets\BaseStack::EVENT_RUN, [Events::class, 'onNotificationInfoWidget']],
+        [AccountTopMenu::class, AccountTopMenu::EVENT_INIT, [Events::class, 'onAccountTopMenuInit']],
     ],
     'consoleControllerMap' => [
-        'firebase' => 'humhub\modules\fcmPush\commands\SendController'
+        'firebase' => 'humhub\modules\fcmPush\commands\SendController',
+    ],
+    'urlManagerRules' => [
+        ['class' => 'humhub\modules\fcmPush\components\UrlRule'],
     ],
 ];
-?>