+++ driver/Fcm.php
@@ -5,7 +5,6 @@ namespace humhub\modules\fcmPush\driver;
 use humhub\modules\fcmPush\components\SendReport;
 use humhub\modules\fcmPush\models\ConfigureForm;
 use humhub\modules\fcmPush\Module;
-use humhub\modules\notification\models\Notification as NotificationHumHub;
 use Kreait\Firebase\Contract\Messaging;
 use Kreait\Firebase\Exception\FirebaseException;
 use Kreait\Firebase\Exception\MessagingException;
@@ -29,12 +28,12 @@ class Fcm implements DriverInterface
         if ($this->messaging === null) {
             Module::registerAutoloader();
 
-            $factory = (new Factory)->withServiceAccount($this->config->getJsonAsArray());
+            $factory = (new Factory())->withServiceAccount($this->config->getJsonAsArray());
             $this->messaging = $factory->createMessaging();
         }
 
         $message = CloudMessage::new()
-            ->withNotification(Notification::create($title, $body))
+            ->withNotification(Notification::create($title, $body, $imageUrl))
             ->withWebPushConfig(['fcm_options' => ['link' => $url]])
             ->withData(['url' => $url, 'notification_count' => $notificationCount]);
 
@@ -65,6 +64,11 @@ class Fcm implements DriverInterface
 
     public function isConfigured(): bool
     {
-        return (!empty($this->config->json) && !empty($this->config->senderId));
+        return !empty($this->config->json) &&
+            !empty($this->config->senderId) &&
+            !empty($this->config->getJsonParam('project_id')) &&
+            !empty($this->config->firebaseApiKey) &&
+            !empty($this->config->firebaseAppId) &&
+            !empty($this->config->firebaseVapidKey);
     }
 }