+++ helpers/MobileAppHelper.php
@@ -2,18 +2,24 @@
 
 namespace humhub\modules\fcmPush\helpers;
 
+use humhub\helpers\DeviceDetectorHelper;
 use Yii;
 use yii\helpers\Json;
 use yii\helpers\Url;
 
 class MobileAppHelper
 {
+    public const SESSION_VAR_SHOW_OPENER = 'mobileAppShowOpener';
+    /**
+     * @deprecated Remove when minimal HumHub mobile app support is v1.0.124 and later
+     */
+    public const SESSION_VAR_HIDE_OPENER = 'mobileAppHideOpener';
+    public const SESSION_VAR_REGISTER_NOTIFICATION = 'mobileAppRegisterNotification';
+    public const SESSION_VAR_UNREGISTER_NOTIFICATION = 'mobileAppUnregisterNotification';
 
-
-    public static function registerLoginScript()
+    public static function registerHideOpenerScript(): void
     {
-
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -23,9 +29,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function registerLogoutScript()
+    public static function registerShowOpenerScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -35,9 +41,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function registerNotificationScript()
+    public static function registerNotificationScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -46,9 +52,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function unregisterNotificationScript()
+    public static function unregisterNotificationScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -57,29 +63,8 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function isAppRequest()
-    {
-        return (
-            (Yii::$app->request->headers->get('x-requested-with', null, true) === 'com.humhub.app') ||
-            (Yii::$app->request->headers->has('x-humhub-app'))
-        );
-    }
-
-    private static function sendFlutterMessage($msg)
+    private static function sendFlutterMessage($msg): void
     {
         Yii::$app->view->registerJs('if (window.flutterChannel) { window.flutterChannel.postMessage(\'' . $msg . '\'); }');
     }
-
-    public static function isIosApp()
-    {
-        $headers = Yii::$app->request->headers;
-
-        if (static::isAppRequest() &&
-            $headers->has('user-agent') &&
-            str_contains($headers->get('user-agent', '', true), 'iPhone')) {
-
-            return true;
-        }
-    }
-
 }