+++ models/ConfigureForm.php
@@ -3,6 +3,8 @@
 namespace humhub\modules\fcmPush\models;
 
 use humhub\modules\fcmPush\Module;
+use humhub\modules\fcmPush\services\WellKnownService;
+use humhub\widgets\Link;
 use Yii;
 use yii\base\InvalidArgumentException;
 use yii\base\Model;
@@ -15,14 +17,19 @@ class ConfigureForm extends Model
     public $humhubInstallId;
 
     public $senderId;
+    public $firebaseApiKey;
+    public $firebaseAppId;
+    public $firebaseVapidKey;
 
     public $json;
-
-    public $serverKey;
 
     public $humhubApiKey;
 
     public $disableAuthChoicesIos;
+
+    public $fileAssetLinks;
+
+    public $fileAppleAssociation;
 
     /**
      * Validate JSON field params
@@ -39,8 +46,7 @@ class ConfigureForm extends Model
             if (isset($arrayCheck[$key])) {
                 if (empty($arrayCheck[$key])) {
                     $errors["empty"] .= $errors["empty"] == "" ? "\"$key\"" : ", \"$key\"";
-                }
-                else {
+                } else {
                     $condition = false;
                     switch ($value['type']) {
                         case "string":
@@ -84,7 +90,9 @@ class ConfigureForm extends Model
         return [
             [['enableEmailGoService', 'disableAuthChoicesIos'], 'boolean'],
             [['senderId'], 'number'],
-            [['serverKey', 'json', 'humhubApiKey'], 'safe'],
+            [['firebaseApiKey', 'firebaseAppId', 'firebaseVapidKey'], 'string'],
+            [['json', 'humhubApiKey'], 'safe'],
+            [['fileAssetLinks', 'fileAppleAssociation'], 'string'],
             ['json', function ($attribute, $params, $validator) {
                 if (empty($this->$attribute)) {
                     return;
@@ -138,18 +146,38 @@ class ConfigureForm extends Model
             'humhubInstallId' => Yii::t('FcmPushModule.base', 'Install ID'),
             'humhubApiKey' => Yii::t('FcmPushModule.base', 'API Key'),
             'senderId' => Yii::t('FcmPushModule.base', 'Sender ID'),
+            'firebaseApiKey' => Yii::t('FcmPushModule.base', 'Web API Key'),
+            'firebaseAppId' => Yii::t('FcmPushModule.base', 'Web App ID'),
+            'firebaseVapidKey' => Yii::t('FcmPushModule.base', 'Key pair of the Web Push certificates'),
             'json' => Yii::t('FcmPushModule.base', 'Service Account (JSON file)'),
-            'serverKey' => Yii::t('FcmPushModule.base', 'Cloud Messaging API (Legacy)'),
             'disableAuthChoicesIos' => Yii::t('FcmPushModule.base', 'Disable AuthChoices on iOS App'),
+            'fileAssetLinks' => Yii::t('FcmPushModule.base', 'Well-known file {fileName}', [
+                'fileName' => '"' . WellKnownService::getFileName('fileAssetLinks') . '"',
+            ]),
+            'fileAppleAssociation' => Yii::t('FcmPushModule.base', 'Well-known file {fileName}', [
+                'fileName' => '"' . WellKnownService::getFileName('fileAppleAssociation') . '"',
+            ]),
         ];
     }
 
     public function attributeHints()
     {
         return [
-            'humhubInstallId' => 'Use this ID to register your API Key.',
-            'serverKey' => 'Please switch to the new "Firebase Cloud Messaging API (V1)" and enter a JSON file in the field above. The old legacy API is only temporarily available for existing installations and is no longer supported or maintained. ',
-            'json' => 'Paste the content of the service account JSON files here. You can find more information in the module instructions.'
+            'humhubInstallId' => Yii::t('FcmPushModule.base', 'Use this ID to register your API Key.'),
+            'firebaseVapidKey' => Yii::t('FcmPushModule.base', 'Firebase Cloud Messaging -> Web Push certificates -> Key pair'),
+            'json' => Yii::t('FcmPushModule.base', 'Paste the content of the service account JSON files here. You can find more information in the module instructions.'),
+            'fileAssetLinks' => Yii::t('FcmPushModule.base', 'URL to the file {fileNameLink}', [
+                'fileNameLink' => Link::to(
+                    WellKnownService::getFileName('fileAssetLinks'),
+                    WellKnownService::getFileRoute('fileAssetLinks'),
+                )->target('_blank'),
+            ]),
+            'fileAppleAssociation' => Yii::t('FcmPushModule.base', 'URL to the file {fileNameLink}', [
+                'fileNameLink' => Link::to(
+                    WellKnownService::getFileName('fileAppleAssociation'),
+                    WellKnownService::getFileRoute('fileAppleAssociation'),
+                )->target('_blank'),
+            ]),
         ];
     }
 
@@ -166,11 +194,14 @@ class ConfigureForm extends Model
         $this->enableEmailGoService = $settings->get('enableEmailGoService', false);
         $this->humhubInstallId = $adminModule->settings->get('installationId');
         $this->senderId = $settings->get('senderId');
+        $this->firebaseApiKey = $settings->get('firebaseApiKey');
+        $this->firebaseAppId = $settings->get('firebaseAppId');
+        $this->firebaseVapidKey = $settings->get('firebaseVapidKey');
         $this->json = $settings->get('json');
-        $this->serverKey = $settings->get('serverKey');
         $this->humhubApiKey = $settings->get('humhubApiKey');
         $this->disableAuthChoicesIos = $settings->get('disableAuthChoicesIos');
-
+        $this->fileAssetLinks = $settings->get('fileAssetLinks');
+        $this->fileAppleAssociation = $settings->get('fileAppleAssociation');
 
         return true;
     }
@@ -182,10 +213,14 @@ class ConfigureForm extends Model
 
         $module->settings->set('enableEmailGoService', $this->enableEmailGoService);
         $module->settings->set('senderId', $this->senderId);
+        $module->settings->set('firebaseApiKey', $this->firebaseApiKey);
+        $module->settings->set('firebaseAppId', $this->firebaseAppId);
+        $module->settings->set('firebaseVapidKey', $this->firebaseVapidKey);
         $module->settings->set('json', $this->json);
-        $module->settings->set('serverKey', $this->serverKey);
         $module->settings->set('humhubApiKey', $this->humhubApiKey);
         $module->settings->set('disableAuthChoicesIos', $this->disableAuthChoicesIos);
+        $module->settings->set('fileAssetLinks', $this->fileAssetLinks);
+        $module->settings->set('fileAppleAssociation', $this->fileAppleAssociation);
 
         return true;
     }
@@ -195,9 +230,14 @@ class ConfigureForm extends Model
         return Json::decode($this->json);
     }
 
+    public function getJsonParam(string $param): ?string
+    {
+        return $this->getJsonAsArray()[$param] ?? null;
+    }
+
     public static function getInstance()
     {
-        $config = new static;
+        $config = new static();
         $config->loadSettings();
 
         return $config;