+++ views/admin/index.php
@@ -1,12 +1,12 @@
 <?php
-/* @var $this \humhub\modules\ui\view\components\View */
 
-/* @var $model \humhub\modules\fcmPush\models\ConfigureForm */
-
-use humhub\widgets\Button;
+use humhub\modules\fcmPush\models\ConfigureForm;
 use humhub\modules\ui\form\widgets\ActiveForm;
+use humhub\modules\ui\icon\widgets\Icon;
+use humhub\widgets\Button;
 use yii\helpers\Html;
 
+/* @var $model ConfigureForm */
 ?>
 <div class="panel panel-default">
     <div class="panel-heading"><?= Yii::t('FcmPushModule.base', '<strong>FireBase Messaging</strong> Configuration'); ?></div>
@@ -16,7 +16,7 @@ use yii\helpers\Html;
         <h3><?= Yii::t('FcmPushModule.base', 'Link Redirection Service') ?></h3>
         <?= $form->field($model, 'enableEmailGoService')->checkbox()
             ->label(Yii::t('FcmPushModule.base', 'Enable Link Redirection Service. In order for links to open in the app on mobile devices, rather than in the mobile browser, all links (e.g. notification emails) need to be routed through the HumHub proxy server. (Experimental Features // <a href="{url}">Privacy Policy</a>)', [
-                'url' => 'https://www.humhub.com/en/privacy/'
+                'url' => 'https://www.humhub.com/en/privacy/',
             ])) ?>
 
         <hr>
@@ -38,17 +38,29 @@ use yii\helpers\Html;
             <?= Yii::t('FcmPushModule.base', 'To send Firebase push notifications with your own Firebase project, enter your access details here.') ?>
         </p>
         <?= Button::info('Installation documentation')->link('https://marketplace.humhub.com/module/fcm-push/installation')->options(['target' => '_blank'])->loader(false) ?>
-        <?= $form->field($model, 'senderId'); ?>
+        <?= $form->field($model, 'senderId') ?>
+        <?= $form->field($model, 'firebaseApiKey') ?>
+        <?= $form->field($model, 'firebaseAppId') ?>
+        <?= $form->field($model, 'firebaseVapidKey') ?>
         <?= $form->field($model, 'json')->textarea(['rows' => 10]); ?>
-        <?php if (!empty($model->serverKey)): ?>
-            <?= $form->field($model, 'serverKey')->textInput(); ?>
-        <?php endif; ?>
         <br/>
 
 
         <?= $form->beginCollapsibleFields('Advanced Settings'); ?>
         <?= $form->field($model, 'disableAuthChoicesIos')->checkbox()
             ->label(Yii::t('FcmPushModule.base', 'Hide third-party login options for app users with iOS.')) ?>
+
+        <?php if (!Yii::$app->urlManager->enablePrettyUrl) : ?>
+        <div class="alert alert-warning">
+            <?= Icon::get('warning') ?>
+            <?= Yii::t('FcmPushModule.base', 'Please enable <a href="{url}" target="_blank">Pretty URLs</a> for proper working of the well-known files.', [
+                'url' => 'https://docs.humhub.org/docs/admin/installation/#pretty-urls',
+            ]) ?>
+        </div>
+        <?php endif; ?>
+
+        <?= $form->field($model, 'fileAssetLinks')->textarea(['rows' => 10]) ?>
+        <?= $form->field($model, 'fileAppleAssociation')->textarea(['rows' => 10]) ?>
         <?= $form->endCollapsibleFields(); ?>
         <br/>
 
@@ -56,12 +68,8 @@ use yii\helpers\Html;
             <?= Html::submitButton(Yii::t('base', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
         </div>
 
-
         <?php ActiveForm::end(); ?>
 
-
-
-
-        <?= Html::a('Mobile App Debug', ['/fcm-push/mobile-app'], ['class' => 'pull-right']); ?>
+        <?= Html::a('Mobile App Debug', ['/fcm-push/admin/mobile-app'], ['class' => 'pull-right']) ?>
     </div>
 </div>